Keep promisses in edge for now
